Fix issue #1
A newline in the input somehow sets !isNan to true, consequently triggering the convertNumber function.  The function returns the variable 'output', which is set to emoji markup based on the number passed.

However, the variable 'output' was prevously defined in a global scope,  and referred to the output textarea. Since the variable wasn't changed to emoji markup based on the numbers 0-9, the function returns 'output' which was previously set to the the textarea element.

The variable used inside the convertNumber function has been changed from 'output' to 'numOutput' to avoid any future confusion.